Move non-data models to value-objects instead
