+++ lib/assets/javascripts/cartodb3/components/modals/add-analysis/analysis-option-models/analysis-option-model.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var camshaftReference = require('../../../../data/camshaft-reference');
-var nodeIds = require('../../../../data/analysis-definition-node-ids');
+var nodeIds = require('../../../../value-objects/analysis-node-ids');
 
 module.exports = cdb.core.Model.extend({
