+++ lib/assets/javascripts/cartodb3/data/layer-definition-model.js
@@ -3,7 +3,7 @@ var _ = require('underscore');
 var syncAbort = require('./backbone/sync-abort');
 var StyleDefinitionModel = require('../editor/style/style-definition-model');
 var StyleCartoCSSModel = require('../editor/style/style-cartocss-model');
-var nodeIds = require('./analysis-definition-node-ids.js');
+var nodeIds = require('../value-objects/analysis-node-ids');
 var camshaftReference = require('./camshaft-reference');
 var layerTypesAndKinds = require('./layer-types-and-kinds');
 var InfowindowModel = require('./infowindow-click-model');
