+++ lib/assets/javascripts/cartodb3/data/layer-definitions-collection.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var Backbone = require('backbone');
 var LayerDefinitionModel = require('./layer-definition-model');
 var layerLetters = require('./layer-letters');
-var nodeIds = require('./analysis-definition-node-ids.js');
+var nodeIds = require('../value-objects/analysis-node-ids');
 var camshaftReference = require('./camshaft-reference');
 var layerTypesAndKinds = require('./layer-types-and-kinds');
