+++ lib/assets/javascripts/cartodb3/data/query-schema-model.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var cdb = require('cartodb.js');
 var Backbone = require('backbone');
 var syncAbort = require('./backbone/sync-abort');
-var geometry = require('./geometry');
+var geometry = require('../value-objects/geometry');
 
 var MAX_GET_LENGTH = 1024;
 var WRAP_SQL_TEMPLATE = 'select * from (<%= sql %>) __wrapped limit <%= rows_sample_size %>';
