+++ lib/assets/javascripts/cartodb3/value-objects/analysis-node-ids.js
