+++ lib/assets/javascripts/cartodb3/value-objects/geometry.js
