+++ lib/assets/test/spec/cartodb3/components/modals/add-analysis/add-analysis-view.spec.js
@@ -1,5 +1,5 @@
 var cdb = require('cartodb.js');
-var geometry = require('../../../../../../javascripts/cartodb3/data/geometry');
+var geometry = require('../../../../../../javascripts/cartodb3/value-objects/geometry');
 var AnalysisDefinitionNodeSourceModel = require('../../../../../../javascripts/cartodb3/data/analysis-definition-node-source-model');
 var AddAnalysisView = require('../../../../../../javascripts/cartodb3/components/modals/add-analysis/add-analysis-view');
