+++ lib/assets/test/spec/cartodb3/editor/layers/layer-content-view/analyses/analysis-source-options-model.spec.js
@@ -1,5 +1,5 @@
 var ConfigModel = require('../../../../../../../javascripts/cartodb3/data/config-model');
-var geometry = require('../../../../../../../javascripts/cartodb3/data/geometry');
+var geometry = require('../../../../../../../javascripts/cartodb3/value-objects/geometry');
 var AnalysisDefinitionNodesCollection = require('../../../../../../../javascripts/cartodb3/data/analysis-definition-nodes-collection');
 var TablesCollection = require('../../../../../../../javascripts/cartodb3/data/tables-collection');
 var AnalysisSourceOptionsModel = require('../../../../../../../javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-source-options-model');
