+++ lib/assets/test/spec/cartodb3/value-objects/analysis-node-ids.spec.js
@@ -1,6 +1,6 @@
-var analysisDefinitionNodeIds = require('../../../../javascripts/cartodb3/data/analysis-definition-node-ids');
+var analysisDefinitionNodeIds = require('../../../../javascripts/cartodb3/value-objects/analysis-node-ids');
 
-describe('data/analysis-definition-node-ids', function () {
+describe('value-objects/analysis-node-ids', function () {
   describe('.next', function () {
     it('should generate a new id for a given source id', function () {
       expect(analysisDefinitionNodeIds.next('a0')).toEqual('a1');
